requirements.txt
PyNaCl
aiohttp>=3.6.0,<3.8.0
async-timeout
attrs
certifi
cffi
chardet
discord.py
idna
multidict
pip
pycparser
requests
setuptools
six
spotipy
typing-extensions
urllib3
yarl
ffmpeg
asyncio
regex
